rename(jwt): Move jwt package outside of util and delete util package
