+++ src/main/java/com/swyp3/babpool/global/config/WebMvcInterceptJwtConfig.java
@@ -1,6 +1,6 @@
 package com.swyp3.babpool.global.config;
 
-import com.swyp3.babpool.global.util.jwt.JwtTokenInterceptor;
+import com.swyp3.babpool.global.jwt.JwtTokenInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
