+++ src/main/java/com/swyp3/babpool/global/jwt/exception/BadCredentialsException.java
@@ -1,6 +1,6 @@
-package com.swyp3.babpool.global.util.jwt.exception;
+package com.swyp3.babpool.global.jwt.exception;
 
-import com.swyp3.babpool.global.util.jwt.exception.errorcode.JwtExceptionErrorCode;
+import com.swyp3.babpool.global.jwt.exception.errorcode.JwtExceptionErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
