+++ src/main/java/com/swyp3/babpool/global/jwt/exception/handler/JwtExceptionHandler.java
@@ -1,8 +1,8 @@
-package com.swyp3.babpool.global.util.jwt.exception.handler;
+package com.swyp3.babpool.global.jwt.exception.handler;
 
 import com.swyp3.babpool.global.common.response.ApiErrorResponse;
-import com.swyp3.babpool.global.util.jwt.exception.BabpoolJwtException;
-import com.swyp3.babpool.global.util.jwt.exception.BadCredentialsException;
+import com.swyp3.babpool.global.jwt.exception.BabpoolJwtException;
+import com.swyp3.babpool.global.jwt.exception.BadCredentialsException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
