+++ src/main/java/com/swyp3/babpool/global/uuid/exception/UuidExceptionHandler.java
@@ -1,7 +1,6 @@
 package com.swyp3.babpool.global.uuid.exception;
 
 import com.swyp3.babpool.global.common.response.ApiErrorResponse;
-import com.swyp3.babpool.global.util.jwt.exception.BabpoolJwtException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
